+++ app/(invites)/[lang]/invite/[inviteId]/page.tsx
@@ -13,14 +13,16 @@ async function getData(inviteId: string): Promise<InvitePreviewData> {
   let domain = `https://${projectId}.web.app`;
 
   const itemsURL = `${domain}/api/v1/invites/${inviteId}/items`;
-  const itemsResponse: Response = await fetch(itemsURL);
+  const itemsResponse: Response = await fetch(itemsURL, {
+    next: { revalidate: 60 },
+  });
   if (!itemsResponse.ok) {
     throw new Error("Failed to fetch items data");
   }
   let items: AlbumItem[] = await itemsResponse.json();
 
   const inviteURL = `${domain}/api/v1/invites/${inviteId}`;
-  const inviteResponse: Response = await fetch(inviteURL);
+  const inviteResponse: Response = await fetch(inviteURL, { next: { revalidate: 60 } });
   if (!inviteResponse.ok) {
     throw new Error("Failed to fetch invite data");
   }