+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Fix #3 for header margins
 - Grid item overlays
+- Revalidate data fetch after 60s
 
 ## 2024-05-29
 - Fix (pinned) item order